pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "A CLI based Chabot with Persistent Memory"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "openai>=1.59.9",
    "prompt-toolkit>=3.0.50",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
    "pymongo>=4.6.0",
]

[project.scripts]
chatbot = "chatbot.chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["chatbot"]